project/requirements.txt
fastapi
uvicorn
psycopg2-binary
pymysql
sqlalchemy
gunicorn
